# WebContent/swagger/properties.json
{
  "swagger": "2.0",
  "info": {
    "title": "AltoroJ REST API",
    "version": "1.0.2"
  },
  "basePath": "/AltoroJ/api",
...
  "paths": {
    "/login": {
      "get": {
        "tags": [
          "1. Login"
        ],
        "summary": "Check if any user is logged in",
        "description": "If a user is loggedin the username will be returned",
        "operationId": "checkLogin",
        "produces": [
          "application/json"
        ],
        "parameters": [
            {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         	}
         ],
        "responses": {
          "401": {
          	"description": "Logged out"	
          },
          "200": {
            "description": "Logged in"
          }
        }
      },
      "post": {
        "tags": [
          "1. Login"
        ],
        "summary": "Login method",
        "description": "After a successful login a token is returned. This is a Bearer token. To authenticate with it use the Authorization header and set value to Bearer empty space and the token value.",
        "operationId": "login",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        	{
            "in": "body",
            "name": "body",
            "description": "Username and password combination to allow users to log-in",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success message when login is complete"
          },
          "400": {
            "description": "Bad parameters: Please check provided values"
          },
          "500": {
            "description": "Internal server error: Please see error message or logs for details"
          }
        }
      }
    },
...
    "/transfer": {
      "post": {
        "tags": [
          "3. Transfer"
        ],
        "description": "Transfer money between two accounts",
        "operationId": "trasnfer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         },
          {
            "in": "body",
            "name": "body",
            "description": "Transfer details including ammount, sender and receiver",
            "required": true,
            "schema": {
              "$ref": "#/definitions/transfer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
           "400": {
            "description": "Bad parameters: Please check provided values"
          },
          "501": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/feedback/submit": {
      "post": {
        "tags": [
          "4. Feedback"
        ],
        "description": "Submit feedback for the bank",
        "operationId": "sendFeedback",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Feedback details including name, email the subject and complete message",
            "required": true,
            "schema": {
              "$ref": "#/definitions/feedback"
            }
          }
        ]
      }
    },
    "/feedback/{feedbackId}": {
      "get": {
        "tags": [
          "4. Feedback"
        ],
        "description": "Retrieve feedback",
        "operationId": "getFeedback",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
           "400": {
            "description": "Bad parameters: Please check provided values"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
         {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         },
          {
            "name": "feedbackId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/logout": {
      "get": {
        "tags": [
          "6. Logout"
        ],
        "description": "Logout from the bank",
        "operationId": "doLogOut",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "500": {
            "description": "Unauthorized request "
          }
        }
      }
    }
  },
  "definitions": {
    "login": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "jsmith"
        },
        "password": {
          "type": "string",
          "example": "demo1234"
        }
      }
    },
...
    "transfer": {
      "type": "object",
      "required": [
        "toAccount",
        "fromAccount",
        "transferAmount"
      ],
      "properties": {
        "toAccount": {
          "type": "string",
          "example": "800002"
        },
        "fromAccount": {
          "type": "string",
          "example": "800003"
        },
        "transferAmount": {
          "type": "string",
          "example": "200"
        }
      }
    },
    "feedback": {
      "type": "object",
      "required": [
        "name",
        "email",
        "subject",
        "message"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "J Smith"
        },
        "email": {
          "type": "string",
          "format": "email",
          "example": "jsmtih@altoromutual.com"
        },
        "subject": {
          "type": "string",
          "example": "Amazing web design"
        },
        "message": {
          "type": "string",
          "example": "I like the new look of your applicaiton"
        }
      }
    },
...
  }
}